:recycle: refactor(#5): 계층별 폴더 구조로 변경
