+++ app/src/main/java/com/umc/phrase/MainActivity.kt
@@ -11,7 +11,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.umc.phrase.commons.ui.theme.PhraseTheme
+import com.umc.phrase.presentation.theme.PhraseTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
