+++ app/src/main/java/com/umc/phrase/data/datasources/UserRemoteDataSource.kt
@@ -1,0 +1,4 @@
+package com.umc.phrase.data.datasources
+
+interface UserRemoteDataSource {
+}
