+++ app/src/main/java/com/umc/phrase/data/model/UserEntity.kt
@@ -1,0 +1,5 @@
+package com.umc.phrase.data.model
+
+data class UserEntity(
+    val name: String,
+)
