+++ app/src/main/java/com/umc/phrase/data/repository/UserRepositoryImpl.kt
@@ -1,0 +1,4 @@
+package com.umc.phrase.data.repository
+
+class UserRepositoryImpl {
+}
