+++ app/src/main/java/com/umc/phrase/domain/model/Book.kt
@@ -1,0 +1,4 @@
+package com.umc.phrase.domain.model
+
+class Book {
+}
