+++ app/src/main/java/com/umc/phrase/domain/model/User.kt
@@ -1,0 +1,5 @@
+package com.umc.phrase.domain.model
+
+data class User (
+    val name: String,
+)
