+++ app/src/main/java/com/umc/phrase/domain/repository/UserRepository.kt
@@ -1,0 +1,4 @@
+package com.umc.phrase.domain.repository
+
+interface UserRepository {
+}
