+++ app/src/main/java/com/umc/phrase/domain/usecase/book/GetBookUseCase.kt
@@ -1,0 +1,4 @@
+package com.umc.phrase.domain.usecase.book
+
+class GetBookUseCase {
+}
