+++ app/src/main/java/com/umc/phrase/domain/usecase/user/GetUserUseCase.kt
@@ -1,0 +1,4 @@
+package com.umc.phrase.domain.usecase.user
+
+class GetUserUseCase {
+}
