+++ app/src/main/java/com/umc/phrase/presentation/components/Button.kt
@@ -1,4 +1,4 @@
-package com.umc.phrase.commons.ui.component
+package com.umc.phrase.presentation.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -25,7 +25,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.umc.phrase.commons.ui.theme.PhraseTheme
+import com.umc.phrase.presentation.theme.PhraseTheme
 
 @Composable
 fun BasicButton(
