+++ app/src/main/java/com/umc/phrase/presentation/components/InputField.kt
@@ -1,4 +1,4 @@
-package com.umc.phrase.commons.ui.component
+package com.umc.phrase.presentation.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -33,7 +33,7 @@ import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.umc.phrase.commons.ui.theme.PhraseTheme
+import com.umc.phrase.presentation.theme.PhraseTheme
 
 enum class BorderType {
     NONE, NORMAL, SUCCESS, ERROR
