+++ app/src/main/java/com/umc/phrase/presentation/screens/community/CommunityScreen.kt
@@ -1,0 +1,4 @@
+package com.umc.phrase.presentation.screens.community
+
+class CommunityScreen {
+}
