+++ app/src/main/java/com/umc/phrase/presentation/screens/home/HomeScreen.kt
@@ -1,0 +1,4 @@
+package com.umc.phrase.presentation.screens.home
+
+class HomeScreen {
+}
