+++ app/src/main/java/com/umc/phrase/presentation/screens/login/LoginScreen.kt
@@ -1,0 +1,4 @@
+package com.umc.phrase.presentation.screens.login
+
+class LoginScreen {
+}
