+++ app/src/main/java/com/umc/phrase/presentation/screens/profile/ProfileScreen.kt
@@ -1,0 +1,4 @@
+package com.umc.phrase.presentation.screens.profile
+
+class ProfileScreen {
+}
