+++ app/src/main/java/com/umc/phrase/presentation/screens/search/SearchScreen.kt
@@ -1,0 +1,4 @@
+package com.umc.phrase.presentation.screens.search
+
+class SearchScreen {
+}
