+++ app/src/main/java/com/umc/phrase/presentation/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.umc.phrase.commons.ui.theme
+package com.umc.phrase.presentation.theme
 
 import androidx.compose.ui.graphics.Color
