+++ app/src/main/java/com/umc/phrase/presentation/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.umc.phrase.commons.ui.theme
+package com.umc.phrase.presentation.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
