+++ app/src/main/java/com/umc/phrase/remote/api/UserApiService.kt
@@ -1,0 +1,4 @@
+package com.umc.phrase.remote.api
+
+interface UserApiService {
+}
