+++ app/src/main/java/com/umc/phrase/remote/datasources/UserRemoteDataSourceImpl.kt
@@ -1,0 +1,4 @@
+package com.umc.phrase.remote.datasources
+
+class UserRemoteDataSourceImpl {
+}
