+++ app/src/main/java/com/umc/phrase/remote/model/UserResponse.kt
@@ -1,0 +1,2 @@
+package com.umc.phrase.remote.model
+
